[ADD] wsite_portal_ext: Correctly manage companies
Remove the usage of street2. The street is stored in the street
attribute. And it found the name of the company in a parent_id if there
is one.

If the partner is a company. It shows information about the company.
Namely, its name, its VAT number, its email, its phone, and its address.

If the partner is a contact of a company. It shows informations about
the company and about the contact. Namely, the name, the email and the
phone of the contact, the name, the VAT and the address of the company.
In fact the VAT and the address must be the same on the contact and on
the company. This is the standard behaviour in Odoo.

So when a partner became a contact of a company it lost his own address
and VAT number. These two field are copied from the company.

When the partner is a person. It shows information about this person.
Namely, his or her name, email, phone, VAT (not required) and address.

It also make the website portal easier to extend.