[FIX] emc_website_portal: PDF download issue
The bug was that trying to download the cooperator certificate or
another PDF in the interface brings an error 500. This was due to a
wrong usage of the function `get_pdf()` and also due to a error of
encoding when using `format()` python function with `{name}`. Instead I
use `"%s" % name` that works great.

To have the `get_pdf()` work correctly, a new definition have to be
added to the class Report.